% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_policy_tree.R
\name{parallel_policy_tree}
\alias{parallel_policy_tree}
\title{Return string \code{"Hello world!"} to R.}
\usage{
parallel_policy_tree(X, Y, depth)
}
\description{
Return string \code{"Hello world!"} to R.
}
