% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_policy_tree.R, R/sparse_policy_tree.R
\name{sparse_policy_tree}
\alias{sparse_policy_tree}
\title{Sparse Policy Tree}
\usage{
sparse_policy_tree(
  X,
  Gamma,
  depth = 2,
  split.step = NULL,
  min.node.size = NULL,
  verbose = NULL
)

sparse_policy_tree(
  X,
  Gamma,
  depth = 2,
  split.step = NULL,
  min.node.size = NULL,
  verbose = NULL
)
}
\arguments{
\item{X}{The covariates used for splitting in the tree (dimension NxP)}

\item{Gamma}{Rewards for each action / treatment (dimension NXD)}

\item{depth}{The number of variables.}

\item{split.step}{consider every n'th split (currently ignored)}

\item{min.node.size}{the smallest node size allowes (currently ignored)}

\item{verbose}{print debug info (currently ignored)
Return string \code{"Hello world!"} to R.}
}
\description{
Sparse Policy Tree
}
