% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_policy_tree.R
\name{sparse_policy_tree}
\alias{sparse_policy_tree}
\title{Return string \code{"Hello world!"} to R.}
\usage{
sparse_policy_tree(X, Gamma, depth)
}
\description{
Return string \code{"Hello world!"} to R.
}
